[HUDI-3581] Reorganize some clazz for hudi flink (#4983)
