+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/async/HoodieAsyncService.java
@@ -170,6 +170,7 @@ public abstract class HoodieAsyncService implements Serializable {
       if (null != callback) {
         callback.apply(null != error);
       }
+      this.started = false;
     });
   }
