+++ hudi-flink/src/main/java/org/apache/hudi/table/format/cow/vector/ParquetDecimalVector.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.table.format.cow;
+package org.apache.hudi.table.format.cow.vector;
 
 import org.apache.flink.table.data.DecimalData;
 import org.apache.flink.table.data.vector.BytesColumnVector;
